% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_proportions.R
\name{dots_to_charlist}
\alias{dots_to_charlist}
\title{Convert dots to a list of character vectors}
\usage{
dots_to_charlist(parent = 1L)
}
\arguments{
\item{call}{a language object reflecting the current call.}

\item{...}{passed from the calling function}
}
\value{
a list of character vectors
}
\description{
This function is intended to allow the user to use NSE within their dot
calls for the express purpose of renaming functions. This function is
intended for internal use to be passed to other MSF packages.
}
\examples{
x <- function(...) {
  dots_to_charlist()
}
x(a = 1, b = TRUE, c = three)
}
\keyword{internal}
