% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fac_from_num}
\alias{fac_from_num}
\title{create factors from numbers}
\usage{
fac_from_num(x)
}
\arguments{
\item{x}{a vector of integers or numerics}
}
\value{
a factor
}
\description{
If the number of unique numbers is five or fewer, then they will simply
be converted to factors in order, otherwise, they will be passed to cut and
pretty, preserving the lowest value.
}
\examples{
fac_from_num(1:100)
fac_from_num(sample(100, 5))
}
