% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epikit-package.R
\docType{package}
\name{epikit-package}
\alias{epikit}
\alias{epikit-package}
\title{epikit: Miscellaneous Tools for the 'R4Epis' Project}
\description{
Contains tools for formatting inline code, renaming redundant
  columns, aggregating age categories, and calculating proportions with
  confidence intervals. This is part of the 'R4Epis' project
  <https://r4epis.netlify.com>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/R4EPI/epikit}
  \item \url{https://r4epis.netlify.com}
  \item \url{https://r4epi.github.io/epikit}
  \item Report bugs at \url{https://github.com/R4EPI/epikit/issues}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})

Authors:
\itemize{
  \item Dirk Schumacher \email{mai@dirk-schumacher.net}
}

Other contributors:
\itemize{
  \item Alex Spina [contributor]
  \item Kate Doyle [contributor]
}

}
\keyword{internal}
